Refactor Entry to abstract class instead of interface
