+++ src/main/java/seedu/address/model/task/Deadline.java
@@ -4,25 +4,33 @@ import java.time.LocalDateTime;
 import java.util.Objects;
 
 import javafx.beans.property.ObjectProperty;
+import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.property.SimpleObjectProperty;
+import javafx.beans.property.SimpleStringProperty;
 import seedu.address.commons.util.CollectionUtil;
 import seedu.address.model.tag.UniqueTagList;
 
-public final class Deadline extends Task{
+public final class Deadline extends Entry{
     protected ObjectProperty<LocalDateTime> deadline;
 
 	public Deadline(Title title, LocalDateTime deadline, UniqueTagList tags, String desc) {
 		this(title, deadline, tags, false, desc);
 	}
 
-	public Deadline(Title title, LocalDateTime deadline, UniqueTagList tags, boolean isMarked, String desc) {
-		super(title, null, tags, isMarked, desc);
-        assert !CollectionUtil.isAnyNull(deadline);
-		this.deadline = new SimpleObjectProperty<>(deadline);
+	public Deadline(Title title, LocalDateTime deadline, UniqueTagList tags, boolean isMarked, String description) {
+        assert !CollectionUtil.isAnyNull(title, tags, description);
+        this.title = new SimpleObjectProperty<>(Title.copy(title));
+        this.tags = new SimpleObjectProperty<>(new UniqueTagList(tags));
+        this.isMarked = new SimpleBooleanProperty(Boolean.valueOf(isMarked));
+        this.description = new SimpleStringProperty(description);
+        if (deadline != null) {
+            this.deadline = new SimpleObjectProperty<>(deadline);
+        } else {
+            this.deadline = new SimpleObjectProperty<>();
+        }
 	}
 	
 	public Deadline(Entry entry) {
-		super(entry);
 		assert entry instanceof Deadline;
 		this.deadline  = new SimpleObjectProperty<>(((Deadline)entry).getDeadline());
 	}
@@ -35,7 +43,6 @@ public final class Deadline extends Task{
 		this.deadline.set(deadline);
 	}
 	
-	@Override
 	public ObjectProperty<LocalDateTime> deadlineObjectProperty() {
 		return deadline;
 	}
