+++ src/main/java/seedu/address/model/task/Entry.java
@@ -1,7 +1,15 @@
 package seedu.address.model.task;
 
+import java.time.LocalDateTime;
+
 import javafx.beans.Observable;
+import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.ObjectProperty;
+import javafx.beans.property.SimpleBooleanProperty;
+import javafx.beans.property.SimpleObjectProperty;
+import javafx.beans.property.SimpleStringProperty;
+import javafx.beans.property.StringProperty;
+import seedu.address.commons.util.CollectionUtil;
 import seedu.address.model.tag.UniqueTagList;
 
 /**
@@ -9,79 +17,139 @@ import seedu.address.model.tag.UniqueTagList;
  * Implementations should guarantee: details are present and not null, field
  * values are validated.
  */
-public interface Entry {
-
+public abstract class Entry {
+	protected ObjectProperty<Title> title;
+    protected ObjectProperty<UniqueTagList> tags;
+    protected BooleanProperty isMarked;
+    protected StringProperty description;
+    
     /**
      * Get the Title for this Entry
      */
-    Title getTitle();
+    public final Title getTitle() {
+        return title.get();
+    }
 
     /**
      * Sets the Title for this Entry
      */
-    void setTitle(Title newTitle);
+    public final void setTitle(Title newTitle) {
+        title.set(newTitle);
+    }
 
     /**
      * Get the titleObjectProperty for this Entry
      */
-    Observable titleObjectProperty();
+    public final ObjectProperty<Title> titleObjectProperty() {
+    	return title;
+    }
 
     /**
      * The returned TagList is a deep copy of the internal TagList,
      * changes on the returned list will not affect the task's internal tags.
      */
-    UniqueTagList getTags();
+    public final UniqueTagList getTags() {
+        return new UniqueTagList(tags.get());
+    }
 
     /**
      * Sets the Tags for this Entry
      */
-    void setTags(UniqueTagList uniqueTagList);
+    public final void setTags(UniqueTagList newTags) {
+        tags.set(new UniqueTagList(newTags));
+    }
 
     /**
      * Add tags for this entry
      */
-    void addTags(UniqueTagList uniqueTagList);
+    public final void addTags(UniqueTagList uniqueTagList) {
+        UniqueTagList updatedTagList = new UniqueTagList(tags.get());
+        updatedTagList.mergeFrom(uniqueTagList);
+        tags.set(updatedTagList);
+    }
 
     /**
      * Remove tags for this entry
      */
-    void removeTags(UniqueTagList tagsToRemove);
+    public final void removeTags(UniqueTagList tagsToRemove) {
+        UniqueTagList updatedTagList = new UniqueTagList(tags.get());
+        updatedTagList.removeFrom(tagsToRemove);
+        tags.set(updatedTagList);
+    }
 
     /**
      * Get the uniqueTagListObjectProperty for this Entry
      */
-    Observable uniqueTagListObjectProperty();
+    public final ObjectProperty<UniqueTagList> uniqueTagListObjectProperty() {
+        return tags;
+    }
 
     /**
      * Get the description for this Entry
      */
-    String getDescription();
+    public final String getDescription() {
+        if (description == null){
+            return "";
+        }
+        return description.get();
+    }
 
     /**
      * Set the description for this Entry
      */
-    void setDescription(String newDescription);
+    public final void setDescription(String newDescription) {
+        if (description == null) {
+            description = new SimpleStringProperty(newDescription);
+            return;
+        }
+        description.set(newDescription);
+    }
 
     /**
      * Get the descriptionProperty for this Entry
      */
-    Observable descriptionProperty();
-
-    /**
-     * Get the deadlineProperty for this Entry
-     */
-    default Observable deadlineObjectProperty() {
-        return null;
+    public final StringProperty descriptionProperty() {
+        return description;
     }
 
     /**
      * Returns true if Entry is marked as completed
      */
-    boolean isMarked();
+    public final boolean isMarked() {
+        return isMarked.get();
+    }
+    
+    /**
+     * Marks the entry as completed.
+     */
+    public final void mark() {
+        this.isMarked.set(true);
+    }
+
+    /**
+     * Unmarks the entry
+     */
+    public final void unmark() {
+        this.isMarked.set(false);
+    }
+
+    /**
+     * Get the isMarkProperty for this Entry
+     */
+    public final Observable isMarkedProperty() {
+        return isMarked;
+    }
+    /**
+     * Returns a string representation of this Entry's marking
+     */
+    public final String markString() {
+        return isMarked() ? "[X] " : "[ ] ";
+    }
+    
     /**
      * Formats the Entry as text, showing all contact details.
      */
-    default String getAsText() {
+    public String getAsText() {
         final StringBuilder builder = new StringBuilder();
         builder.append(getTitle());
         if (!getTags().isEmpty()) {
@@ -98,7 +166,7 @@ public interface Entry {
     /**
      * Returns a string representation of this Entry's tags
      */
-    default String tagsString() {
+    public final String tagsString() {
         final StringBuffer buffer = new StringBuffer();
         final String separator = ", ";
         getTags().forEach(tag -> buffer.append(tag).append(separator));
@@ -108,27 +176,4 @@ public interface Entry {
             return buffer.substring(0, buffer.length() - separator.length());
         }
     }
-
-
-
-    /**
-     * Marks the entry as completed.
-     */
-    void mark();
-
-    /**
-     * Unmarks the entry
-     */
-    void unmark();
-
-    /**
-     * Get the isMarkProperty for this Entry
-     */
-    Observable isMarkedProperty();
-
-    /**
-     * Returns a string representation of this Entry's marking
-     */
-    String markString();
-
 }
