+++ src/main/java/seedu/address/model/task/UniqueTaskList.java
@@ -40,7 +40,12 @@ public class UniqueTaskList implements Iterable<Entry> {
             new Callback<Entry, Observable[]>() {
         @Override
         public Observable[] call(Entry entry) {
-            return new Observable[] { entry.titleObjectProperty(), entry.uniqueTagListObjectProperty(), entry.deadlineObjectProperty(), entry.descriptionProperty(), entry.isMarkedProperty()};
+        	if (entry instanceof Task) {
+        		return new Observable[] { entry.titleObjectProperty(), entry.uniqueTagListObjectProperty(), ((Task)entry).deadlineObjectProperty(), entry.descriptionProperty(), entry.isMarkedProperty()};
+        	} else {
+        		return new Observable[] { entry.titleObjectProperty(), entry.uniqueTagListObjectProperty(), entry.descriptionProperty(), entry.isMarkedProperty()};
+        	}
+            
             /*if (entry instanceof Deadline) {
                 return new Observable[] { entry.titleObjectProperty(), entry.uniqueTagListObjectProperty(), entry.deadlineObjectProperty(), entry.descriptionProperty()};
             } else {
